Conflicts:
	krita/plugins/extensions/dockers/CMakeLists.txt

+++ kexi/core/kexiproject.cpp
@@ -1,6 +1,6 @@
 /* This file is part of the KDE project
    Copyright (C) 2003 Lucijan Busch <lucijan@kde.org>
-   Copyright (C) 2003-2010 Jarosław Staniek <staniek@kde.org>
+   Copyright (C) 2003-2012 Jarosław Staniek <staniek@kde.org>
 
    This library is free software; you can redistribute it and/or
    modify it under the terms of the GNU Library General Public
@@ -18,9 +18,9 @@
  * Boston, MA 02110-1301, USA.
 */
 
-#include <qfile.h>
-#include <qapplication.h>
-#include <qdom.h>
+#include <QFile>
+#include <QApplication>
+#include <QDomDocument>
 
 #include <kmimetype.h>
 #include <kdebug.h>
@@ -28,14 +28,14 @@
 
 #include <kexiutils/identifier.h>
 
-#include <kexidb/connection.h>
-#include <kexidb/cursor.h>
-#include <kexidb/driver.h>
-#include <kexidb/drivermanager.h>
-#include <kexidb/utils.h>
-#include <kexidb/parser/parser.h>
-#include <kexidb/msghandler.h>
-#include <kexidb/dbproperties.h>
+#include <db/connection.h>
+#include <db/cursor.h>
+#include <db/driver.h>
+#include <db/drivermanager.h>
+#include <db/utils.h>
+#include <db/parser/parser.h>
+#include <db/msghandler.h>
+#include <db/dbproperties.h>
 #include <kexiutils/utils.h>
 
 #include "kexiproject.h"
@@ -99,6 +99,16 @@ public:
             classNames.insert(id, partClass);
         }
 //! @todo what to do with extra ids for the same part or extra class name for the same ID?
+    }
+
+    //! @return user name for the current project
+    //! @todo the name is taken from connection but it also can be specified otherwise
+    //!       if the same connection data is shared by multiple users. This will be especially
+    //!       true for 3-tier architectures.
+    QString userName() const
+    {
+        QString name = connection->data()->userName;
+        return name.isNull() ? "" : name;
     }
 
     QPointer<KexiDB::Connection> connection;
@@ -461,7 +471,7 @@ bool KexiProject::createInternalStructures(bool insideTransaction)
         if (!d->connection->isReadOnly()) {
             partsTableOk = d->connection->createTable(t_parts, true/*replaceExisting*/);
 
-            KexiDB::FieldList *fl = t_parts->subList("p_id", "p_name", "p_mime", "p_url");
+            QScopedPointer<KexiDB::FieldList> fl(t_parts->subList("p_id", "p_name", "p_mime", "p_url"));
 #define INSERT_RECORD(id, groupName, name) \
             if (partsTableOk) { \
                 partsTableOk = d->connection->insertRecord(*fl, QVariant(int(KexiPart::id)), \
@@ -486,6 +496,24 @@ bool KexiProject::createInternalStructures(bool insideTransaction)
     if (!partsTableOk) {
         delete t_parts;
         return false;
+    }
+
+    // User data storage
+    KexiDB::InternalTableSchema *t_userdata = new KexiDB::InternalTableSchema("kexi__userdata");
+    t_userdata->addField(new KexiDB::Field("d_user", KexiDB::Field::Text, KexiDB::Field::NotNull))
+        .addField(new KexiDB::Field("o_id", KexiDB::Field::Integer, KexiDB::Field::Unsigned | KexiDB::Field::NotNull))
+        .addField(new KexiDB::Field("d_sub_id", KexiDB::Field::Text, KexiDB::Field::NotNull | KexiDB::Field::NotEmpty))
+        .addField(new KexiDB::Field("d_data", KexiDB::Field::LongText));
+
+    bool containsKexi__userdataTable = d->connection->drv_containsTable("kexi__userdata");
+    if (containsKexi__userdataTable) {
+        d->connection->insertInternalTable(*t_userdata);
+    }
+    else if (!d->connection->isReadOnly()) {
+        if (!d->connection->createTable(t_userdata, true/*replaceExisting*/)) {
+            delete t_userdata;
+            return false;
+        }
     }
 
     if (insideTransaction) {
@@ -644,7 +672,7 @@ bool KexiProject::retrieveItems()
         int ident = cursor->value(0).toInt(&ok);
         QString objName(cursor->value(1).toString());
         if (ok && (ident > 0) && !d->connection->isInternalTableSchema(objName)
-                && KexiUtils::isIdentifier(objName))
+                && KexiDB::isIdentifier(objName))
         {
             KexiPart::Item *it = new KexiPart::Item();
             it->setIdentifier(ident);
@@ -865,6 +893,10 @@ bool KexiProject::removeObject(KexiPart::Item& item)
         }
         if (!d->connection->removeObject(item.identifier())) {
             setError(d->connection);
+            return false;
+        }
+        if (!removeUserDataBlock(item.identifier())) {
+            setError(ERR_DELETE_SERVER_ERROR, i18n("Could not remove object's user data."));
             return false;
         }
         if (!tg.commit()) {
@@ -1097,45 +1129,48 @@ bool KexiProject::checkProject(const QString& singlePartClass)
         setError(d->connection);
         return false;
     }
-    QString sql("SELECT p_id, p_name, p_mime, p_url FROM kexi__parts ORDER BY p_id");
-    if (!singlePartClass.isEmpty()) {
-        sql.append(QString(" WHERE p_url=%1").arg(d->connection->driver()->escapeString(singlePartClass)));
-    }
-    KexiDB::Cursor *cursor = d->connection->executeQuery(sql);
-    if (!cursor) {
-        setError(d->connection);
-        return false;
-    }
+    bool containsKexi__partsTable = d->connection->drv_containsTable("kexi__parts");
+    if (containsKexi__partsTable) { // check if kexi__parts exists, if missing, createInternalStructures() will create it
+        QString sql("SELECT p_id, p_name, p_mime, p_url FROM kexi__parts ORDER BY p_id");
+        if (!singlePartClass.isEmpty()) {
+            sql.append(QString(" WHERE p_url=%1").arg(d->connection->driver()->escapeString(singlePartClass)));
+        }
+        KexiDB::Cursor *cursor = d->connection->executeQuery(sql);
+        if (!cursor) {
+            setError(d->connection);
+            return false;
+        }
 
-    bool saved = false;
-    for (cursor->moveFirst(); !cursor->eof(); cursor->moveNext()) {
-        const QString partMime( cursor->value(2).toString() );
-        QString partClass( cursor->value(3).toString() );
-        partClass = realPartClass(partClass, partMime);
-        if (partClass == QLatin1String("uk.co.piggz.report")) { // compatibility
-            partClass = QLatin1String("org.kexi-project.report");
+        bool saved = false;
+        for (cursor->moveFirst(); !cursor->eof(); cursor->moveNext()) {
+            const QString partMime( cursor->value(2).toString() );
+            QString partClass( cursor->value(3).toString() );
+            partClass = realPartClass(partClass, partMime);
+            if (partClass == QLatin1String("uk.co.piggz.report")) { // compatibility
+                partClass = QLatin1String("org.kexi-project.report");
+            }
+            KexiPart::Info *info = Kexi::partManager().infoForClass(partClass);
+            bool ok;
+            const int classId = cursor->value(0).toInt(&ok);
+            if (!ok || classId <= 0) {
+                kWarning() << "Invalid class Id" << classId << "; part" << partClass << "will not be used";
+            }
+            if (info && ok && classId > 0) {
+                d->saveClassId(partClass, classId);
+                saved = true;
+            }
+            else {
+                KexiPart::MissingPart m;
+                m.name = cursor->value(1).toString();
+                m.className = partClass;
+                d->missingParts.append(m);
+            }
         }
-        KexiPart::Info *info = Kexi::partManager().infoForClass(partClass);
-        bool ok;
-        const int classId = cursor->value(0).toInt(&ok);
-        if (!ok || classId <= 0) {
-            kWarning() << "Invalid class Id" << classId << "; part" << partClass << "will not be used";
-        }
-        if (info && ok && classId > 0) {
-            d->saveClassId(partClass, classId);
-            saved = true;
-        }
-        else {
-            KexiPart::MissingPart m;
-            m.name = cursor->value(1).toString();
-            m.className = partClass;
-            d->missingParts.append(m);
-        }
-    }
 
-    d->connection->deleteCursor(cursor);
-    if (!saved && !singlePartClass.isEmpty()) {
-        return false; // failure is single part class was not found
+        d->connection->deleteCursor(cursor);
+        if (!saved && !singlePartClass.isEmpty()) {
+            return false; // failure is single part class was not found
+        }
     }
     return true;
 }
@@ -1174,7 +1209,7 @@ bool KexiProject::createIdForPart(const KexiPart::Info& info)
         d->connection->tableSchema("kexi__parts");
     if (!ts)
         return false;
-    KexiDB::FieldList *fl = ts->subList("p_id", "p_name", "p_mime", "p_url");
+    QScopedPointer<KexiDB::FieldList> fl(ts->subList("p_id", "p_name", "p_mime", "p_url"));
     //kDebug() << "fieldlist: " << (fl ? fl->debugString() : QString());
     if (!fl)
         return false;
@@ -1206,6 +1241,72 @@ bool KexiProject::createIdForPart(const KexiPart::Info& info)
 KexiPart::MissingPartsList KexiProject::missingParts() const
 {
     return d->missingParts;
+}
+
+static bool checkObjectId(const char* method, int objectID)
+{
+    if (objectID <= 0) {
+        kWarning() << method <<  ": Invalid objectID" << objectID;
+        return false;
+    }
+    return true;
+}
+
+tristate KexiProject::loadUserDataBlock(int objectID, const QString& dataID, QString *dataString)
+{
+    if (!checkObjectId("loadUserDataBlock", objectID)) {
+        return false;
+    }
+    return d->connection->querySingleString(
+               QString::fromLatin1("SELECT d_data FROM kexi__userdata WHERE o_id=") + QString::number(objectID)
+                + " AND " + KexiDB::sqlWhere(d->connection->driver(), KexiDB::Field::Text, "d_user", d->userName())
+                + " AND " + KexiDB::sqlWhere(d->connection->driver(), KexiDB::Field::Text, "d_sub_id", dataID),
+               *dataString);
+}
+
+bool KexiProject::storeUserDataBlock(int objectID, const QString& dataID, const QString &dataString)
+{
+    if (!checkObjectId("storeUserDataBlock", objectID)) {
+        return false;
+    }
+    QString sql(QString::fromLatin1(
+                    "SELECT kexi__userdata.o_id FROM kexi__userdata WHERE o_id=%1").arg(objectID));
+    QString sql_sub(
+        KexiDB::sqlWhere(d->connection->driver(), KexiDB::Field::Text, "d_user", d->userName())
+        + " AND " + KexiDB::sqlWhere(d->connection->driver(), KexiDB::Field::Text, "d_sub_id", dataID));
+
+    bool ok;
+    bool exists = d->connection->resultExists(sql + " AND " + sql_sub, ok);
+    if (!ok)
+        return false;
+    if (exists) {
+        return d->connection->executeSQL("UPDATE kexi__userdata SET d_data="
+                          + d->connection->driver()->valueToSQL(KexiDB::Field::LongText, dataString)
+                          + " WHERE o_id=" + QString::number(objectID) + " AND " + sql_sub);
+    }
+    return d->connection->executeSQL(
+               QString::fromLatin1("INSERT INTO kexi__userdata (d_user, o_id, d_sub_id, d_data) VALUES (")
+               + d->connection->driver()->valueToSQL(KexiDB::Field::Text, d->userName())
+               + ", " + QString::number(objectID)
+               + ", " + d->connection->driver()->valueToSQL(KexiDB::Field::Text, dataID)
+               + ", " + d->connection->driver()->valueToSQL(KexiDB::Field::LongText, dataString)
+               + ")");
+}
+
+bool KexiProject::removeUserDataBlock(int objectID, const QString& dataID)
+{
+    if (!checkObjectId("removeUserDataBlock", objectID)) {
+        return false;
+    }
+    if (dataID.isEmpty())
+        return KexiDB::deleteRow(*d->connection, "kexi__userdata",
+                                 "o_id", KexiDB::Field::Integer, objectID,
+                                 "d_user", KexiDB::Field::Text, d->userName());
+    else
+        return KexiDB::deleteRow(*d->connection, "kexi__userdata",
+                                 "o_id", KexiDB::Field::Integer, objectID,
+                                 "d_user", KexiDB::Field::Text, d->userName(),
+                                 "d_sub_id", KexiDB::Field::Text, dataID);
 }
 
 #include "kexiproject.moc"
